move auth utility functions into the auth module
Also, cleanup import order in some of the modules. =)
